fix: cleanup after failed download
This should help in cases where the binary download somehow fails.
Removing all the assets for that specific version should allow
coc-tabnine to try to redownload once it is restarted. This solution is
far from fail-safe though as there is currently no way to validate, via
a hash or similar, if a binary is valid or not.

Related issues: #32, #26, #24, #6.